fixes #137 - switch ssize_t usage to int

+++ tests/test069.cpp
@@ -20,10 +20,10 @@ int main()
   {
     rapidcsv::Document doc(path);
 
-    unittest::ExpectEqual(ssize_t, doc.GetRowIdx("-"), -1);
-    unittest::ExpectEqual(ssize_t, doc.GetRowIdx("1"), -1);
-    unittest::ExpectEqual(ssize_t, doc.GetRowIdx("2"), -1);
-    unittest::ExpectEqual(ssize_t, doc.GetRowIdx("3"), -1);
+    unittest::ExpectEqual(int, doc.GetRowIdx("-"), -1);
+    unittest::ExpectEqual(int, doc.GetRowIdx("1"), -1);
+    unittest::ExpectEqual(int, doc.GetRowIdx("2"), -1);
+    unittest::ExpectEqual(int, doc.GetRowIdx("3"), -1);
   }
   catch (const std::exception& ex)
   {
